+++ AccountActionsHistory.js
@@ -15,10 +15,16 @@ import { formatDateTime, getFullName } from './util';
 
 class AccountActionsHistory extends React.Component {
   static manifest = Object.freeze({
+    account: {
+      type: 'okapi',
+      records: 'accounts',
+      path: 'accounts?query=id=%{activeRecord.accountId}&limit=1',
+    },
     accountActions: {
       type: 'okapi',
       records: 'feefineactions',
-      path: 'feefineactions?query=(accountId=%{activeRecord.accountId})',
+      accumulate: 'true',
+      path: 'feefineactions?query=(accountId=%{activeRecord.accountId})&limit=50',
     },
     activeRecord: {
       accountId: '0',
@@ -28,6 +34,7 @@ class AccountActionsHistory extends React.Component {
   static propTypes = {
     stripes: PropTypes.object.isRequired,
     resources: PropTypes.shape({
+      intl: PropTypes.object.isRequired,
       accountActions: PropTypes.object,
     }),
     mutator: PropTypes.shape({
@@ -37,10 +44,16 @@ class AccountActionsHistory extends React.Component {
       feefineactions: PropTypes.shape({
         POST: PropTypes.func.isRequired,
       }),
+      accountActions: PropTypes.shape({
+        GET: PropTypes.func.isRequired,
+      }),
     }),
     account: PropTypes.object,
     user: PropTypes.object,
+    history: PropTypes.object,
+    patronGroup: PropTypes.object,
     onCancel: PropTypes.func.isRequired,
+    onClickViewLoanActionsHistory: PropTypes.func.isRequired,
   };
   constructor(props) {
     super(props);
@@ -50,7 +63,6 @@ class AccountActionsHistory extends React.Component {
     this.connectedActions = props.stripes.connect(Actions);
     this.error = this.error.bind(this);
     this.comment = this.comment.bind(this);
-    this.accounts = [this.props.account];
 
     const { stripes } = props;
 
@@ -66,11 +78,14 @@ class AccountActionsHistory extends React.Component {
     };
 
     this.state = {
+      data: [],
       actions: {
         pay: false,
         cancellation: false,
+        waive: false,
         waiveModal: false,
         comment: false,
+        regular: false,
       },
       checkedAccounts: {},
       sortOrder: [
@@ -91,9 +106,20 @@ class AccountActionsHistory extends React.Component {
   }
 
   componentDidMount() {
-    const { account } = this.props;
-    const accountId = (account) ? account.id : '';
-    this.props.mutator.activeRecord.update({ accountId });
+    const { history } = this.props;
+    const str = history.location.search || '';
+    const n = str.indexOf('account=');
+    const id = str.substring(n + 8, n + 44);
+    this.props.mutator.activeRecord.update({ accountId: id });
+    this.getAccountActions();
+  }
+
+  getAccountActions = () => {
+    return this.props.mutator.accountActions.GET().then(records => {
+      this.setState({
+        data: records,
+      });
+    });
   }
 
   onChangeActions(actions) {
@@ -135,9 +161,11 @@ class AccountActionsHistory extends React.Component {
 
   render() {
     const { sortOrder, sortDirection } = this.state;
-    const { onCancel, account, stripes } = this.props;
+    const { onCancel, stripes } = this.props;
+    const account = _.get(this.props.resources, ['account', 'records', 0]) || this.props.account;
 
     const user = this.props.user;
+    const patron = this.props.patronGroup;
 
     const columnMapping = {
       Comments: (<span>{this.props.stripes.intl.formatMessage({ id: 'ui-users.details.columns.comments' })}<Button style={{ float: 'right', marginLeft: '50px' }} onClick={this.comment}>+ New</Button></span>),
@@ -146,17 +174,21 @@ class AccountActionsHistory extends React.Component {
     const accountActionsFormatter = {
       // Action: aa => loanActionMap[la.action],
       'Action Date': action => formatDateTime(action.dateAction, stripes.locale),
-      'Action': action => action.typeAction,
-      'Amount': action => action.amountAction || '-',
-      'Balance': action => action.balance || '-',
-      'Transaction number': action => action.transactionNumber,
+      'Action': action => action.typeAction + (action.paymentMethod ? ('-' + action.paymentMethod) : ' '),
+      'Amount': action => (action.amountAction > 0 ? parseFloat(action.amountAction).toFixed(2) : '-'),
+      'Balance': action => (action.balance > 0 ? parseFloat(action.balance).toFixed(2) : '-'),
+      'Transaction information': action => action.transactionNumber || '-',
       'Created at': action => action.createdAt,
       'Source': action => action.source,
       'Comments': action => action.comments,
     };
 
-    const actions = _.get(this.props.resources, ['accountActions', 'records'], []);
+    const actions = this.state.data || [];
     const actionsSort = _.orderBy(actions, [this.sortMap[sortOrder[0]], this.sortMap[sortOrder[1]]], sortDirection);
+    const amount = (account.amount) ? parseFloat(account.amount).toFixed(2) : '-';
+    const remaining = (account.remaining) ? parseFloat(account.remaining).toFixed(2) : '0.00';
+    const loanId = account.loanId || '';
+    const disabled = (_.get(account, ['status', 'name'], '') === 'Closed');
 
     return (
       <Paneset isRoot>
@@ -165,54 +197,75 @@ class AccountActionsHistory extends React.Component {
           defaultWidth="100%"
           dismissible
           onClose={onCancel}
-          paneTitle={<Link to={`/users/view/${user.id}`}>{getFullName(user)}</Link>}
+          paneTitle={`Fees/Fines - ${getFullName(user)} (${_.upperFirst(patron.group)}) `}
         >
           <Row>
             <Col xs={12}>
-              <Button buttonStyle="primary" onClick={this.pay}>Pay</Button>
-              <Button buttonStyle="primary" onClick={this.waive}>Waive</Button>
-              <Button buttonStyle="primary">Refund</Button>
-              <Button buttonStyle="primary">Transfer</Button>
-              <Button buttonStyle="primary" onClick={this.error}>Error</Button>
-              <img alt="" src="https://png.icons8.com/ios/25/666666/upload.png" />
+              <Button disabled={disabled} buttonStyle="primary" onClick={this.pay}>Pay</Button>
+              <Button disabled={disabled} buttonStyle="primary" onClick={this.waive}>Waive</Button>
+              <Button disabled buttonStyle="primary">Refund</Button>
+              <Button disabled buttonStyle="primary">Transfer</Button>
+              <Button disabled={disabled} buttonStyle="primary" onClick={this.error}>Error</Button>
             </Col>
           </Row>
 
           <Row>
-            <Col xs={2} >
-              <KeyValue label="Fee/fine type" value={_.get(account, ['feeFineType'], '')} />
+            <Col xs={1.5} >
+              <KeyValue label={this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.feetype' })} value={_.get(account, ['feeFineType'], '-')} />
             </Col>
-            <Col xs={2} >
-              <KeyValue label="Fee/fine owner" value={_.get(account, ['feeFineOwner'], '')} />
+            <Col xs={1.5} >
+              <KeyValue label={this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.owner' })} value={_.get(account, ['feeFineOwner'], '-')} />
             </Col>
-            <Col xs={2} >
-              <KeyValue label="Charge date" value={_.get(account, ['dateCreated'], '')} />
+            <Col xs={1.5} >
+              <KeyValue label={this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.billedate' })} value={_.get(account, ['metadata', 'createdDate']) ? formatDateTime(_.get(account, ['metadata', 'createdDate'])) : '-'} />
             </Col>
-            <Col xs={2} >
-              <KeyValue label="Original fee/fine" value={_.get(account, ['charged'], '')} />
+            <Col xs={1.5} >
+              <KeyValue label={this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.billedamount' })} value={amount} />
             </Col>
-            <Col xs={2} >
-              <KeyValue label="Outstanding amount" value={_.get(account, ['remaining'], '')} />
+            <Col xs={1.5} >
+              <KeyValue label={this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.remainingamount' })} value={remaining} />
             </Col>
-            <Col xs={2} >
-              <KeyValue label="Loan details" value="" />
+            <Col xs={1.5} >
+              <KeyValue label={this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.latest' })} value={_.get(account, ['paymentStatus', 'name'], '-')} />
+            </Col>
+            <Col xs={1.5} >
+              {(loanId !== 0) ?
+                <KeyValue
+                  label="Loan details"
+                  value={
+                    <button onClick={(e) => { this.props.onClickViewLoanActionsHistory(e, { id: loanId }); }}>
+                      {this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.loan' })}
+                    </button>}
+                />
+                :
+                <KeyValue
+                  label="Loan details"
+                  value={this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.loan' })}
+                />
+              }
             </Col>
           </Row>
           <Row>
-            <Col xs={2} >
-              <KeyValue label="Item type" value={_.get(account, ['itemType'], '')} />
+            <Col xs={1.5} >
+              <KeyValue label={this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.instance' })} value={_.get(account, ['title'], '-')} />
             </Col>
-            <Col xs={2} >
-              <KeyValue label="Instance" value={_.get(account, ['item'], '')} />
+            <Col xs={1.5} >
+              <KeyValue label={this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.type' })} value={_.get(account, ['materialType'], '-')} />
             </Col>
-            <Col xs={2} >
-              <KeyValue label="Barcode (falta el ID)" value={<Link to={`/inventory/view/${_.get(account, ['itemId'], '')}?query=${_.get(account, ['itemId'], '')}`}>{_.get(account, ['barcode'], '')}</Link>} />
+            <Col xs={1.5} >
+              <KeyValue label={this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.barcode' })} value={<Link to={`/inventory/view/${_.get(account, ['itemId'], '')}?query=${_.get(account, ['itemId'], '')}`}>{_.get(account, ['barcode'], '-')}</Link>} />
             </Col>
-            <Col xs={2} >
-              <KeyValue label="Call number" value={_.get(account, ['callNumber'], '')} />
+            <Col xs={1.5} >
+              <KeyValue label={this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.callnumber' })} value={_.get(account, ['callNumber'], '-')} />
             </Col>
-            <Col xs={2} >
-              <KeyValue label="Location" value="" />
+            <Col xs={1.5} >
+              <KeyValue label={this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.location' })} value={_.get(account, ['location'], '-')} />
+            </Col>
+            <Col xs={1.5} >
+              <KeyValue label={this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.duedate' })} value={formatDateTime(account.dueDate) || '-'} />
+            </Col>
+            <Col xs={1.5} >
+              <KeyValue label={this.props.stripes.intl.formatMessage({ id: 'ui-users.details.field.returnedate' })} value={formatDateTime(account.returnedDate) || '-'} />
             </Col>
           </Row>
           <br />
@@ -220,21 +273,23 @@ class AccountActionsHistory extends React.Component {
             id="list-accountactions"
             formatter={accountActionsFormatter}
             columnMapping={columnMapping}
-            visibleColumns={['Action Date', 'Action', 'Amount', 'Balance', 'Transaction number', 'Created at', 'Source', 'Comments']}
-            contentData={actionsSort}
+            visibleColumns={['Action Date', 'Action', 'Amount', 'Balance', 'Transaction information', 'Created at', 'Source', 'Comments']}
+            contentData={(account.id === (actions[0] || {}).accountId) ? actionsSort : []}
             fullWidth
             onHeaderClick={this.onSort}
             sortOrder={sortOrder[0]}
             sortDirection={`${sortDirection[0]}ending`}
-            columnWidths={{ 'Action': 100, 'Amount': 100, 'Balance': 100, 'Transaction number': 100, 'Created at': 100, 'Source': 200, 'Comments': 700 }}
+            columnWidths={{ 'Action': 250, 'Amount': 100, 'Balance': 100, 'Transaction information': 200, 'Created at': 100, 'Source': 200, 'Comments': 700 }}
           />
-
           <this.connectedActions
             actions={this.state.actions}
             onChangeActions={this.onChangeActions}
             user={this.props.user}
-            accounts={this.accounts}
+            balance={account.remaining || 0}
+            accounts={[account]}
+            handleEdit={this.getAccountActions}
           />
+
         </Pane>
       </Paneset>
     );